Demonstration of primary-foreign-key assumption
The crux of the assumption is that when resolving a relation, say
`Foo.bazz`, the primary key of the `Foo` is used as the identifier
yet the cache key construction expects to be given the primary
key of the `Bazz` model there.

There are two tests which this commit adds -- both demonstrate
similar ways in which this can fail.